test_example3/project/lollms_core/pyproject.toml
[project]
name = "lollms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "ascii-colors>=0.1.3",
    "autopep8>=2.3.2",
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "python-socketio[asyncio-client,client]>=5.13.0",
    "pyyaml>=6.0.2",
    "requests",
    "safe-store",
    "setuptools",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "wget>=3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
